\name{stderr_to_file}
\alias{stderr_to_file}
\title{stderr_to_file}
\usage{
  stderr_to_file(filename, open = "at", ...)
}
\arguments{
  \item{filename}{output file}

  \item{open}{file open mode; see \code{\link{file}}}

  \item{...}{arguments passed to \code{\link{file}()}}
}
\description{
  Set stderr messages to specific file
}

