\name{stdout_to_file}
\alias{stdout_to_file}
\title{stdout_to_file}
\usage{
  stdout_to_file(filename, open = "at", ...)
}
\arguments{
  \item{filename}{output file}

  \item{open}{file open mode; see \code{\link{file}}}

  \item{...}{arguments passed to \code{\link{file}()}}
}
\description{
  Set stdout messages to specific file
}

